Fix: resolve all -Wsign-conversion 'changes value' sentinels (Bucket 6)
Hand-audited each of the 13 sign-flip-to-all-ones warnings; all are
intentional sentinels or type mismatches, none are missing error checks:

dkim-test.c -- dkim_test_dns_get() returns size_t but used bare -1 as
the 'nothing queued' sentinel.  The sole caller (dkim-keys.c) already
compares against (size_t)-1 with a comment explaining the idiom.
Change all nine return sites to return (size_t)-1 to make the intent
explicit and match the call-site convention.

dkim.c:4067 -- sig->sig_flags &= ~DKIM_SIGFLAG_PASSED: standard
bit-clear on a u_int field; ~(int 0x04) = int -5 promoting to u_int.
Cast the flag to u_int before inverting: ~(u_int)DKIM_SIGFLAG_PASSED.

opendkim-db.c:507 -- db_flags &= ~DKIMF_DB_FLAG_MAKELOCK: same
bit-clear pattern (flag 0x20, ~0x20 = -33 as int).
Fix: ~(u_int)DKIMF_DB_FLAG_MAKELOCK.

opendkim-db.h:28 + opendkim-db.c -- DKIMF_DB_TYPE_UNKNOWN is (-1)
but db_type was u_int; dkimf_db_type() already returns int and the
type-code table uses int.  Change db_type from u_int to int in the
private struct; remove the (u_int) workaround cast from the comparison
at db.c:491.

Found via -Wsign-conversion under strict build
(-DOPENDKIM_ENABLE_STRICT_C=ON -DOPENDKIM_ENABLE_EXTRA_WARNINGS=ON).

+++ opendkim/opendkim-db.c
@@ -112,7 +112,7 @@ struct dkimf_db
 {
 	u_int			db_flags;
 	u_int			db_iflags;
-	u_int			db_type;
+	int			db_type;
 	int			db_status;
 	int			db_nrecs;
 	pthread_mutex_t *	db_lock;
@@ -488,7 +488,7 @@ dkimf_db_open(DKIMF_DB *db, char *name, u_int flags, pthread_mutex_t *lock,
 				new->db_type = dbtypes[c].code;
 		}
 
-		if (new->db_type == (u_int) DKIMF_DB_TYPE_UNKNOWN)
+		if (new->db_type == DKIMF_DB_TYPE_UNKNOWN)
 		{
 			free(new);
 			if (err != NULL)
@@ -504,7 +504,7 @@ dkimf_db_open(DKIMF_DB *db, char *name, u_int flags, pthread_mutex_t *lock,
 	if (lock != NULL)
 	{
 		new->db_lock = lock;
-		new->db_flags &= ~DKIMF_DB_FLAG_MAKELOCK;
+		new->db_flags &= ~(u_int)DKIMF_DB_FLAG_MAKELOCK;
 	}
 	else if ((new->db_flags & DKIMF_DB_FLAG_MAKELOCK) != 0)
 	{
